+++ releases.md
@@ -1,8 +1,12 @@
 # Releases
 
+## Unreleased
+
+  - Use queue for `newobj` too to avoid invoking user code during object allocation.
+
 ## v1.1.2
 
-    - Fix handling of GC compaction (I hope).
+  - Fix handling of GC compaction (I hope).
 
 ## v0.1.0